+++ src/pages/LoginPage.tsx
@@ -1,36 +1,9 @@
-import React, { FormEvent } from "react";
+import React, { ReactElement } from "react";
 import styled from "styled-components";
-import { Button } from "components/common/Button";
-import Input from "components/common/Input";
-import { setUser } from "stores/user";
-import { useDispatch } from "react-redux";
 import { DanceImg, LogoImg } from "assets/images";
-import { useNavigate } from "react-router-dom";
-import useInput from "hooks/useInput";
-import { DEFAULT_PROFILE_IMAGE, Path } from "utils/constants";
+import LoginForm from "components/auth/LoginForm";
 
-function LoginPage() {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const [name, onInputName] = useInput("");
-
-  const onEnter = (e: FormEvent) => {
-    e.preventDefault();
-
-    if (name.trim().length === 0) {
-      alert("이름을 입력해주세요");
-      return;
-    }
-
-    const body = {
-      userName: name,
-      userId: Math.random(),
-      profileImage: DEFAULT_PROFILE_IMAGE,
-    };
-    dispatch(setUser(body));
-    navigate(Path.Chat);
-  };
-
+function LoginPage(): ReactElement {
   return (
     <Container>
       <Content>
@@ -41,17 +14,7 @@ function LoginPage() {
           Work OS 올인원 협업툴
         </h1>
         <StyledImg src={DanceImg} alt="dance-image" />
-        <form onSubmit={onEnter}>
-          <StyledInput
-            width={"260px"}
-            height={"40px"}
-            onChange={onInputName}
-            placeholder="이름을 입력하세요"
-          />
-          <Button variant={"secondary"} width={"260px"}>
-            입장하기
-          </Button>
-        </form>
+        <LoginForm />
       </Content>
     </Container>
   );
@@ -91,8 +54,4 @@ const StyledImg = styled.img`
   border-radius: 15px;
   width: 260px;
   margin-bottom: 30px;
-`;
-
-const StyledInput = styled(Input)`
-  margin-bottom: 20px;
 `;
