+++ src/styles/styled.d.ts
@@ -1,10 +1,9 @@
-import 'styled-components';
-import { ColorsTypes, FontSizeTypes } from './theme';
+import "styled-components";
+import { ColorsTypes, FontSizeTypes } from "./theme";
 
-declare module 'styled-components' {
+declare module "styled-components" {
   export interface DefaultTheme {
     colors: ColorsTypes;
-
     fontSize: FontSizeTypes;
   }
 }
